fixing shader stuff, images are still broken, but there's the loading spinner, you can't use u32 in the shader uniforms

+++ src/renderer/mtl/shaders.metal
@@ -56,8 +56,8 @@ struct Uniforms {
     float feather;
     float strokeMult;
     float strokeThr;
-    int texType;
-    int shaderType;
+    float texType;
+    float shaderType;
     float hasMask;
     float padding[19];
 };
